+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Change how relationship schemas are loaded so they are valid as Orbit.js models
+
 ## [0.6.1] - 2022-08-31
 
 ### Added