+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Autofocus the text search input when searching for assets
 - Improve the styling of the asset visual carousel so the controls are easier to see
+- Improve how asset photos are loaded so the 'file--file' entity can come from the local cache if available
 
 ### Fixed